backup-sidecar/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
azure-common==1.1.28
    # via azure-mgmt-storage
azure-core==1.32.0
    # via
    #   azure-mgmt-core
    #   azure-storage-blob
azure-mgmt-core==1.5.0
    # via azure-mgmt-storage
azure-mgmt-storage==22.0.0
    # via backup-sidecar (pyproject.toml)
azure-storage-blob==12.24.1
    # via backup-sidecar (pyproject.toml)
certifi==2025.1.31
    # via requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.1
    # via requests
cryptography==44.0.0
    # via azure-storage-blob
idna==3.10
    # via requests
isodate==0.7.2
    # via
    #   azure-mgmt-storage
    #   azure-storage-blob
pycparser==2.22
    # via cffi
pyyaml==6.0.2
    # via backup-sidecar (pyproject.toml)
requests==2.32.3
    # via azure-core
six==1.17.0
    # via azure-core
typing-extensions==4.12.2
    # via
    #   azure-core
    #   azure-mgmt-storage
    #   azure-storage-blob
urllib3==2.3.0
    # via requests
